requirements_sail.txt
accelerate==1.10.1
bitsandbytes==0.47.0
cuml-cu13==25.12.0
cupy-cuda13x==13.6.0
ipywidgets==8.1.7
lightning==2.5.5
matplotlib==3.10.6
nibabel==5.3.2
notebook==7.4.5
numpy==2.2
pandas==2.3.2
Pillow==12.0.0
psutil==7.2.1
pyarrow==22.0.0
pyright==1.1.408
PyYAML==6.0.3
scikit-image==0.26.0
scikit-learn==1.7.2
scipy==1.13.1
shap==0.50.0
SimpleITK==2.5.3
timm==1.0.22
transformers==4.56.0